Bug 1435715 - No other action should be allowed when re-arranging tab… (#3663)
* Bug 1435715 - No other action should be allowed when re-arranging tabs. Firefox will crash.

* Don't move tabs that are being inserted or deleted.

* Ok, don't crash for real this time

* Cancel the drop operation if the tabs changed out from under our feet.

* Addressed review comment about performBatchUpdates block.

+++ Client/Frontend/Browser/TopTabsViewController.swift
@@ -37,6 +37,7 @@ class TopTabsViewController: UIViewController {
     let tabManager: TabManager
     weak var delegate: TopTabsDelegate?
     fileprivate var isPrivate = false
+    fileprivate var isDragging = false
 
     lazy var collectionView: UICollectionView = {
         let collectionView = UICollectionView(frame: CGRect.zero, collectionViewLayout: TopTabsViewLayout())
@@ -327,7 +328,18 @@ extension TopTabsViewController: UICollectionViewDataSource {
 
 @available(iOS 11.0, *)
 extension TopTabsViewController: UICollectionViewDragDelegate {
+    func collectionView(_ collectionView: UICollectionView, dragSessionWillBegin session: UIDragSession) {
+        isDragging = true
+    }
+
+    func collectionView(_ collectionView: UICollectionView, dragSessionDidEnd session: UIDragSession) {
+        isDragging = false
+    }
+
     func collectionView(_ collectionView: UICollectionView, itemsForBeginning session: UIDragSession, at indexPath: IndexPath) -> [UIDragItem] {
+        // We need to store the earliest oldTabs. So if one already exists use that.
+        self.oldTabs = self.oldTabs ?? tabStore
+
         let tab = tabStore[indexPath.item]
 
         // Get the tab's current URL. If it is `nil`, check the `sessionData` since
@@ -364,18 +376,23 @@ extension TopTabsViewController: UICollectionViewDropDelegate {
             return
         }
 
-        collectionView.performBatchUpdates({
-            self.tabManager.moveTab(isPrivate: self.isPrivate, fromIndex: sourceIndex, toIndex: destinationIndexPath.item)
-            self.tabStore = self.tabsToDisplay
-            collectionView.moveItem(at: IndexPath(item: sourceIndex, section: 0), to: destinationIndexPath)
-        })
-
         coordinator.drop(dragItem, toItemAt: destinationIndexPath)
+        isDragging = false
+
+        self.tabManager.moveTab(isPrivate: self.isPrivate, fromIndex: sourceIndex, toIndex: destinationIndexPath.item)
+        self.performTabUpdates()
     }
 
     func collectionView(_ collectionView: UICollectionView, dropSessionDidUpdate session: UIDropSession, withDestinationIndexPath destinationIndexPath: IndexPath?) -> UICollectionViewDropProposal {
-        guard let _ = session.localDragSession else {
+        guard let localDragSession = session.localDragSession, let item = localDragSession.items.first, let tab = item.localObject as? Tab else {
             return UICollectionViewDropProposal(operation: .forbidden)
+        }
+
+        // If the `isDragging` is not `true` by the time we get here, we've had other
+        // add/remove operations happen while the drag was going on. We must return a
+        // `.cancel` operation continuously until `isDragging` can be reset.
+        guard tabStore.index(of: tab) != nil, isDragging else {
+            return UICollectionViewDropProposal(operation: .cancel)
         }
 
         return UICollectionViewDropProposal(operation: .move, intent: .insertAtDestinationIndexPath)
@@ -404,22 +421,38 @@ extension TopTabsViewController: TabEventHandler {
 
 // Collection Diff (animations)
 extension TopTabsViewController {
+    struct TopTabMoveChange: Hashable {
+        let from: IndexPath
+        let to: IndexPath
+
+        var hashValue: Int {
+            return from.hashValue + to.hashValue
+        }
+
+        // Consider equality when from/to are equal as well as swapped. This is because
+        // moving a tab from index 2 to index 1 will result in TWO changes: 2 -> 1 and 1 -> 2
+        // We only need to keep *one* of those two changes when dealing with a move.
+        static func ==(lhs: TopTabsViewController.TopTabMoveChange, rhs: TopTabsViewController.TopTabMoveChange) -> Bool {
+            return (lhs.from == rhs.from && lhs.to == rhs.to) || (lhs.from == rhs.to && lhs.to == rhs.from)
+        }
+    }
 
     struct TopTabChangeSet {
         let reloads: Set<IndexPath>
         let inserts: Set<IndexPath>
         let deletes: Set<IndexPath>
+        let moves: Set<TopTabMoveChange>
 
-        init(reloadArr: [IndexPath], insertArr: [IndexPath], deleteArr: [IndexPath]) {
+        init(reloadArr: [IndexPath], insertArr: [IndexPath], deleteArr: [IndexPath], moveArr: [TopTabMoveChange]) {
             reloads = Set(reloadArr)
             inserts = Set(insertArr)
             deletes = Set(deleteArr)
+            moves = Set(moveArr)
         }
 
-        var all: [Set<IndexPath>] {
-            return [inserts, reloads, deletes]
+        var isEmpty: Bool {
+            return reloads.isEmpty && inserts.isEmpty && deletes.isEmpty && moves.isEmpty
         }
-
     }
 
     // create a TopTabChangeSet which is a snapshot of updates to perfrom on a collectionView
@@ -438,20 +471,27 @@ extension TopTabsViewController {
             return nil
         }
 
+        let moves: [TopTabMoveChange] = newTabs.enumerated().flatMap { newIndex, tab in
+            if let oldIndex = oldTabs.index(of: tab), oldIndex != newIndex {
+                return TopTabMoveChange(from: IndexPath(row: oldIndex, section: 0), to: IndexPath(row: newIndex, section: 0))
+            }
+            return nil
+        }
+
         // Create based on what is visibile but filter out tabs we are about to insert/delete.
         let reloads: [IndexPath] = reloadTabs.flatMap { tab in
             guard let tab = tab, newTabs.index(of: tab) != nil else {
                 return nil
             }
             return IndexPath(row: newTabs.index(of: tab)!, section: 0)
-            }.filter { return inserts.index(of: $0) == nil && deletes.index(of: $0) == nil }
+        }.filter { return inserts.index(of: $0) == nil && deletes.index(of: $0) == nil }
 
-        return TopTabChangeSet(reloadArr: reloads, insertArr: inserts, deleteArr: deletes)
+        return TopTabChangeSet(reloadArr: reloads, insertArr: inserts, deleteArr: deletes, moveArr: moves)
     }
 
     func updateTabsFrom(_ oldTabs: [Tab]?, to newTabs: [Tab], on completion: (() -> Void)? = nil) {
         assertIsMainThread("Updates can only be performed from the main thread")
-        guard let oldTabs = oldTabs, !self.isUpdating, !self.pendingReloadData else {
+        guard let oldTabs = oldTabs, !self.isUpdating, !self.pendingReloadData, !self.isDragging else {
             return
         }
 
@@ -460,7 +500,7 @@ extension TopTabsViewController {
         flushPendingChanges()
 
         // If there are no changes. We have nothing to do
-        if update.all.every({ $0.isEmpty }) {
+        if update.isEmpty {
             completion?()
             return
         }
@@ -468,19 +508,25 @@ extension TopTabsViewController {
         // The actual update block. We update the dataStore right before we do the UI updates.
         let updateBlock = {
             self.tabStore = newTabs
-            self.collectionView.deleteItems(at: Array(update.deletes))
-            self.collectionView.insertItems(at: Array(update.inserts))
-            self.collectionView.reloadItems(at: Array(update.reloads))
+
+            // Only consider moves if no other operations are pending.
+            if update.deletes.count == 0, update.inserts.count == 0, update.reloads.count == 0 {
+                for move in update.moves {
+                    self.collectionView.moveItem(at: move.from, to: move.to)
+                }
+            } else {
+                self.collectionView.deleteItems(at: Array(update.deletes))
+                self.collectionView.insertItems(at: Array(update.inserts))
+                self.collectionView.reloadItems(at: Array(update.reloads))
+            }
         }
 
         //Lets lock any other updates from happening.
         self.isUpdating = true
+        self.isDragging = false
         self.pendingUpdatesToTabs = newTabs // This var helps other mutations that might happen while updating.
 
-        // The actual update
-        UIView.animate(withDuration: TopTabsUX.AnimationSpeed, animations: {
-            self.collectionView.performBatchUpdates(updateBlock)
-        }) { (_) in
+        let onComplete: () -> Void = {
             self.isUpdating = false
             self.pendingUpdatesToTabs = []
             // Sometimes there might be a pending reload. Lets do that.
@@ -495,6 +541,20 @@ extension TopTabsViewController {
                     self.scrollToCurrentTab()
                 }
             })
+        }
+
+        // The actual update. Only animate the changes if no tabs have moved
+        // as a result of drag-and-drop.
+        if update.moves.count == 0 {
+            UIView.animate(withDuration: TopTabsUX.AnimationSpeed, animations: {
+                self.collectionView.performBatchUpdates(updateBlock)
+            }) { (_) in
+                onComplete()
+            }
+        } else {
+            self.collectionView.performBatchUpdates(updateBlock) { _ in
+                onComplete()
+            }
         }
     }
 
@@ -512,6 +572,7 @@ extension TopTabsViewController {
         }
 
         isUpdating = true
+        isDragging = false
         self.tabStore = self.tabsToDisplay
         self.newTab.isUserInteractionEnabled = false
         self.flushPendingChanges()
